requirements.txt
# Core ML/AI dependencies
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Music processing
music21>=9.1.0
pretty_midi>=0.2.9
mido>=1.2.10
librosa>=0.10.0

# Data processing
pandas>=1.3.0
h5py>=3.7.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment tracking
tensorboard>=2.10.0
wandb>=0.13.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.10.0

# Development
pre-commit>=2.20.0
jupyter>=1.0.0
ipython>=8.0.0

# Utilities
tqdm>=4.64.0
click>=8.0.0
colorama>=0.4.0
rich>=13.0.0